BUGFIX: now works with fractions of seconds ie 3.5 sec BUG: when time < 1/freq then no sound at all

+++ player.py
@@ -13,10 +13,11 @@ import pulseaudio as pa
 #        player.write(out.astype(numpy.int16).tostring())
 #    player.drain()
 
+#BUG, when time is shorter than 1/freq then there is no sound
 def playtone(freq, amplitude, time):
     tab = []
     with pa.simple.open(direction=pa.STREAM_PLAYBACK, format=pa.SAMPLE_S16LE, rate=44100, channels=1) as player:
-        for x in range(0,freq*time):
+        for x in range(0,int(freq*time)):
             tab = numpy.sin(freq / 44100.0 * 2 * numpy.pi * numpy.array(range(0,44100/freq)))*amplitude
             player.write(tab.astype(numpy.int16).tostring())
         player.drain()
